fix semaphores, fix exception recovery, dynamic memory calculation, memory manager info

+++ Kernel/include/bounded_queue.h
@@ -17,6 +17,6 @@ void init_queue(queue_t* q);
 uint8_t is_full(queue_t* q);
 uint8_t is_empty(queue_t* q);
 uint8_t enqueue(queue_t* q, int value);
-uint8_t dequeue(queue_t* q, int* value);
+uint8_t dequeue(queue_t* q, int64_t* value);
 
 #endif // QUEUE_H
